+++ internal/validation/gateway/httproute_test.go
@@ -1,14 +1,17 @@
 package gateway
 
 import (
+	"context"
 	"fmt"
 	"testing"
 
-	"github.com/kong/go-kong/kong"
+	"github.com/blang/semver/v4"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
+
+	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/parser"
 )
 
 func TestValidateHTTPRoute(t *testing.T) {
@@ -16,8 +19,6 @@ func TestValidateHTTPRoute(t *testing.T) {
 		nonexistentListener = gatewayv1beta1.SectionName("listener-that-doesnt-exist")
 		group               = gatewayv1beta1.Group("gateway.networking.k8s.io")
 		defaultGWNamespace  = gatewayv1beta1.Namespace(corev1.NamespaceDefault)
-		pathMatchRegex      = gatewayv1beta1.PathMatchRegularExpression
-		headerMatchRegex    = gatewayv1beta1.HeaderMatchRegularExpression
 		exampleGroup        = gatewayv1beta1.Group("example")
 		podKind             = gatewayv1beta1.Kind("Pod")
 	)
@@ -256,113 +257,6 @@ func TestValidateHTTPRoute(t *testing.T) {
 			err:           fmt.Errorf("queryparam matching is not yet supported for httproute"),
 		},
 		{
-			msg: "if an HTTPRoute is using regex path matching it fails validation due to lack of support",
-			route: &gatewayv1beta1.HTTPRoute{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: corev1.NamespaceDefault,
-					Name:      "testing-httproute",
-				},
-				Spec: gatewayv1beta1.HTTPRouteSpec{
-					CommonRouteSpec: gatewayv1beta1.CommonRouteSpec{
-						ParentRefs: []gatewayv1beta1.ParentReference{{
-							Name: "testing-gateway",
-						}},
-					},
-					Rules: []gatewayv1beta1.HTTPRouteRule{{
-						Matches: []gatewayv1beta1.HTTPRouteMatch{{
-							Path: &gatewayv1beta1.HTTPPathMatch{
-								Type:  &pathMatchRegex,
-								Value: kong.String("^path/to/stuff/*$"),
-							},
-						}},
-						BackendRefs: []gatewayv1beta1.HTTPBackendRef{{
-							BackendRef: gatewayv1beta1.BackendRef{
-								BackendObjectReference: gatewayv1beta1.BackendObjectReference{
-									Namespace: &defaultGWNamespace,
-								},
-							},
-						}},
-					}},
-				},
-			},
-			gateways: []*gatewayv1beta1.Gateway{{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: corev1.NamespaceDefault,
-					Name:      "testing-gateway",
-				},
-				Spec: gatewayv1beta1.GatewaySpec{
-					Listeners: []gatewayv1beta1.Listener{{
-						Name:     "http",
-						Port:     80,
-						Protocol: (gatewayv1beta1.HTTPProtocolType),
-						AllowedRoutes: &gatewayv1beta1.AllowedRoutes{
-							Kinds: []gatewayv1beta1.RouteGroupKind{{
-								Group: &group,
-								Kind:  "HTTPRoute",
-							}},
-						},
-					}},
-				},
-			}},
-			valid:         false,
-			validationMsg: "httproute spec did not pass validation",
-			err:           fmt.Errorf("regex path matching is not yet supported for httproute"),
-		},
-		{
-			msg: "if an HTTPRoute is using regex header matching it fails validation due to lack of support",
-			route: &gatewayv1beta1.HTTPRoute{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: corev1.NamespaceDefault,
-					Name:      "testing-httproute",
-				},
-				Spec: gatewayv1beta1.HTTPRouteSpec{
-					CommonRouteSpec: gatewayv1beta1.CommonRouteSpec{
-						ParentRefs: []gatewayv1beta1.ParentReference{{
-							Name: "testing-gateway",
-						}},
-					},
-					Rules: []gatewayv1beta1.HTTPRouteRule{{
-						Matches: []gatewayv1beta1.HTTPRouteMatch{{
-							Headers: []gatewayv1beta1.HTTPHeaderMatch{{
-								Type:  &headerMatchRegex,
-								Name:  "Content-Type",
-								Value: "audio/vorbis",
-							}},
-						}},
-						BackendRefs: []gatewayv1beta1.HTTPBackendRef{{
-							BackendRef: gatewayv1beta1.BackendRef{
-								BackendObjectReference: gatewayv1beta1.BackendObjectReference{
-									Namespace: &defaultGWNamespace,
-								},
-							},
-						}},
-					}},
-				},
-			},
-			gateways: []*gatewayv1beta1.Gateway{{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: corev1.NamespaceDefault,
-					Name:      "testing-gateway",
-				},
-				Spec: gatewayv1beta1.GatewaySpec{
-					Listeners: []gatewayv1beta1.Listener{{
-						Name:     "http",
-						Port:     80,
-						Protocol: (gatewayv1beta1.HTTPProtocolType),
-						AllowedRoutes: &gatewayv1beta1.AllowedRoutes{
-							Kinds: []gatewayv1beta1.RouteGroupKind{{
-								Group: &group,
-								Kind:  "HTTPRoute",
-							}},
-						},
-					}},
-				},
-			}},
-			valid:         false,
-			validationMsg: "httproute spec did not pass validation",
-			err:           fmt.Errorf("regex header matching is not yet supported for httproute"),
-		},
-		{
 			msg: "we don't support any group except core kubernetes for backendRefs",
 			route: &gatewayv1beta1.HTTPRoute{
 				ObjectMeta: metav1.ObjectMeta{
@@ -478,7 +372,10 @@ func TestValidateHTTPRoute(t *testing.T) {
 			err:           fmt.Errorf("Pod is not a supported kind for httproute backendRefs, only Service is supported"),
 		},
 	} {
-		valid, validMsg, err := ValidateHTTPRoute(tt.route, tt.gateways...)
+		// Passed Kong version is irrelevant for the above test cases.
+		valid, validMsg, err := ValidateHTTPRoute(
+			context.Background(), nil, parser.FeatureFlags{}, semver.MustParse("3.0.0"), tt.route, tt.gateways...,
+		)
 		assert.Equal(t, tt.valid, valid, tt.msg)
 		assert.Equal(t, tt.validationMsg, validMsg, tt.msg)
 		assert.Equal(t, tt.err, err, tt.msg)